+++ src/HashRouter.tsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react';
 // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
 // @ts-ignore
-import { createBrowserHistory } from 'history';
+import { createHashHistory } from 'history';
 
 import { Router } from './Router';
 import { History } from './types';
@@ -11,14 +11,14 @@ type Props = {
   timeout?: number;
 };
 
-export const BrowserRouter: React.FC<Props> = ({
+export const HashRouter: React.FC<Props> = ({
   window,
   timeout,
   children,
 }) => {
   const history = useRef<History>();
   if (!history.current) {
-    history.current = createBrowserHistory({ window });
+    history.current = createHashHistory({ window });
   }
   return (
     <Router history={history.current} timeout={timeout}>
@@ -27,4 +27,4 @@ export const BrowserRouter: React.FC<Props> = ({
   );
 };
 
-export default BrowserRouter;
+export default HashRouter;